requirements.txt
fastapi
gunicorn
uvicorn[standard]
pandas
beautifulsoup4
langchain
langchain-pinecone
langchain-google-genai
langchain-community
pypdf
pinecone-client
